New mix files + updates
Ran mix.new.

Added a driver file which ports the terminal and listens for keystrokes.

Added a game file which manages some of the logic.

Added a pieces file which manages new piece creation, piece
initialisation.

Added a supervisor file which will keep the driver alive, as well as
the game.

Added a formatter file which I’ll implement to display relevant output
to the terminal.